Revert ":zap: :arrow_right: 11009번 BigInteger -> BigDecimal 로 전환 시도"
This reverts commit 4f6826bffdafbec6cab4341b00586403a0bff062.

+++ questions/src/main/java/baekjoon/processing/Main11009.java
@@ -1,7 +1,7 @@
 package baekjoon.processing;
 
 import java.io.*;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.StringTokenizer;
 
 public class Main11009 {
@@ -25,13 +25,13 @@ public class Main11009 {
             final int RED_BALLS = Integer.parseInt(st.nextToken());
             final int WHITE_BALLS = Integer.parseInt(st.nextToken());
 
-            BigDecimal[] fraction = null; //매 차례별 확률 누적값
-            BigDecimal[] winningFraction = new BigDecimal[2]; //빨간공 뽑기 직전까지의 확률(다음차례에 재활용)
+            BigInteger[] fraction = null; //매 차례별 확률 누적값
+            BigInteger[] winningFraction = new BigInteger[2]; //빨간공 뽑기 직전까지의 확률(다음차례에 재활용)
             /*
                 이전에 계산한 확률 누적값 재사용하기 위한 변수들
             */
-            winningFraction[NUMERATOR] = BigDecimal.ONE;
-            winningFraction[DENOMINATOR] = BigDecimal.ONE;
+            winningFraction[NUMERATOR] = BigInteger.ONE;
+            winningFraction[DENOMINATOR] = BigInteger.ONE;
             /*
                 dreamoon이 매 차례마다 빨간 공 뽑을 확률
                 (1번째, 3번째, 5번째... 에 뽑을 확률)
@@ -39,8 +39,8 @@ public class Main11009 {
             for(int i=1, loop=-1; i<=(WHITE_BALLS + 1); i+=2, loop++) {
                 boolean usingSavedValue = loop > 0;
 
-                BigDecimal numerator = winningFraction[NUMERATOR];
-                BigDecimal denominator = winningFraction[DENOMINATOR];
+                BigInteger numerator = winningFraction[NUMERATOR];
+                BigInteger denominator = winningFraction[DENOMINATOR];
 
                 int whiteBallLoopValue = usingSavedValue ? WHITE_BALLS-(loop+1) : WHITE_BALLS;
                 int redBallLoopValue = RED_BALLS;
@@ -51,29 +51,29 @@ public class Main11009 {
                     흰 공 뽑는 경우의 수 (j < i)
                 */
                 for(int j=startIndex; j<i; j++) {
-                    denominator = denominator.multiply(BigDecimal.valueOf(redBallLoopValue + whiteBallLoopValue));
-                    numerator = numerator.multiply(BigDecimal.valueOf(whiteBallLoopValue--));
+                    denominator = denominator.multiply(BigInteger.valueOf(redBallLoopValue + whiteBallLoopValue));
+                    numerator = numerator.multiply(BigInteger.valueOf(whiteBallLoopValue--));
                 }
                 //흰 공 뽑는 확률 저장 (+ 약분)
-                BigDecimal gcd = getGCD(numerator, denominator);
+                BigInteger gcd = getGCD(numerator, denominator);
                 winningFraction[NUMERATOR] = (numerator = numerator.divide(gcd));
                 winningFraction[DENOMINATOR] = (denominator = denominator.divide(gcd));
                 /*
                     [마지막]
                     빨간 공 뽑는 경우의 수 (j = i)
                 */
-                denominator = denominator.multiply(BigDecimal.valueOf(redBallLoopValue + whiteBallLoopValue));
-                numerator = numerator.multiply(BigDecimal.valueOf(redBallLoopValue));
+                denominator = denominator.multiply(BigInteger.valueOf(redBallLoopValue + whiteBallLoopValue));
+                numerator = numerator.multiply(BigInteger.valueOf(redBallLoopValue));
                 //약분하기
                 gcd = getGCD(numerator, denominator);
                 numerator = numerator.divide(gcd);
                 denominator = denominator.divide(gcd);
                 //지난차례와 이번차례의 확률 덧셈
-                if(fraction == null) fraction = new BigDecimal[] {numerator, denominator};
+                if(fraction == null) fraction = new BigInteger[] {numerator, denominator};
                 else fraction = addFractions(numerator, denominator, fraction[NUMERATOR], fraction[DENOMINATOR]);
             }
             //약분하기
-            BigDecimal gcd = getGCD(fraction[NUMERATOR], fraction[DENOMINATOR]);
+            BigInteger gcd = getGCD(fraction[NUMERATOR], fraction[DENOMINATOR]);
             sb.append(fraction[NUMERATOR].divide(gcd)).append('/').append(fraction[DENOMINATOR].divide(gcd)).append("\n");
         }
         bw.write(sb.toString());
@@ -88,9 +88,9 @@ public class Main11009 {
      * @param b
      * @return
      */
-    public static BigDecimal getGCD(BigDecimal a, BigDecimal b) {
-        BigDecimal remainder;
-        while((a.remainder(b).compareTo(BigDecimal.ZERO) != 0)) {
+    public static BigInteger getGCD(BigInteger a, BigInteger b) {
+        BigInteger remainder;
+        while((a.remainder(b).compareTo(BigInteger.ZERO) != 0)) {
             remainder = a.remainder(b);
             a = b;
             b = remainder;
@@ -106,10 +106,10 @@ public class Main11009 {
      * @param b_denominator b 분모
      * @return a분자/a분모 + b분자/b분모
      */
-    public static BigDecimal[] addFractions(BigDecimal a_numerator, BigDecimal a_denominator, BigDecimal b_numerator, BigDecimal b_denominator) {
-        BigDecimal[] result = new BigDecimal[2];
-        BigDecimal gcd = getGCD(a_denominator, b_denominator);
-        BigDecimal lcm = a_denominator.multiply(b_denominator).divide(gcd);
+    public static BigInteger[] addFractions(BigInteger a_numerator, BigInteger a_denominator, BigInteger b_numerator, BigInteger b_denominator) {
+        BigInteger[] result = new BigInteger[2];
+        BigInteger gcd = getGCD(a_denominator, b_denominator);
+        BigInteger lcm = a_denominator.multiply(b_denominator).divide(gcd);
         result[NUMERATOR] = (lcm.divide(a_denominator).multiply(a_numerator)).add(lcm.divide(b_denominator).multiply(b_numerator));
         result[DENOMINATOR] = lcm;
         //약분하기
